refactor(types): split custom-fields.ts into 3 per-entity files
