+++ packages/types/src/encryption.ts
@@ -1,10 +1,11 @@
 import type { AuditEventType, AuditActor, AuditLogEntry } from "./audit-log.js";
 import type { KdfMasterKey } from "./crypto-keys.js";
-import type { FieldDefinition, FieldValue, FieldType } from "./custom-fields.js";
 import type { AcknowledgementRequest } from "./entities/acknowledgement.js";
 import type { BoardMessage } from "./entities/board-message.js";
 import type { Channel } from "./entities/channel.js";
 import type { CustomFront } from "./entities/custom-front.js";
+import type { FieldDefinition, FieldType } from "./entities/field-definition.js";
+import type { FieldValue } from "./entities/field-value.js";
 import type { FrontingComment } from "./entities/fronting-comment.js";
 import type { FrontingSession } from "./entities/fronting-session.js";
 import type { InnerWorldEntity } from "./entities/innerworld-entity.js";
